refactor: 동문 수첩 검색 컴포넌트 상태 관리 및 비즈니스 로직 커스텀 훅 분리
