+++ apps/web/src/features/alumni-contacts/model/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useAlumniContactsSearchInput } from './useAlumniContactsSearchInput';
