+++ apps/web/src/features/alumni-contacts/model/index.ts
@@ -1,0 +1,1 @@
+export { useAlumniContactsSearchInput } from './hooks';
